requirements.txt
Flask==2.2.3
Jinja2==3.0.1
mysql-connector-python==8.0.32
